requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-telegram-bot==20.6
alembic==1.12.1
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6